Split out unit tests to separate files.
